+++ src/index.js
@@ -43,7 +43,7 @@ export { default as ListHeader } from './components/ListView/ListHeader.vue'
 export { default as ListRows } from './components/ListView/ListRows.vue'
 export { default as ListRow } from './components/ListView/ListRow.vue'
 export { default as ListRowItem } from './components/ListView/ListRowItem.vue'
-export { default as SelectBanner } from './components/ListView/SelectBanner.vue'
+export { default as ListSelectBanner } from './components/ListView/ListSelectBanner.vue'
 export { default as Toast } from './components/Toast.vue'
 export { toast, Toasts } from './components/toast.js'
 export { default as Tooltip } from './components/Tooltip.vue'
